[#12] feat : seasons - summer, delete others
- create summer component
- delete other component (animations...)
