+++ src/lib/index.tsx
@@ -1,5 +1,1 @@
-export { default as RippleAnimation } from "./RippleAnimation";
-export { default as FlowerAnimation } from "./FlowerAnimation";
-export { default as FallenLeavesAnimation } from "./FallenLeavesAnimation";
-export { default as SnowAnimation } from "./SnowAnimation";
-export { default as RainAnimation } from "./RainAnimation";
+export { default as Summer } from "./seasons/summer";
